Enable the new scanner, fixed a packet length bug
